repo-add: parse options and flags in arguments until "--" is reached
utils::get_current_cmdname is called right at the start of main, because
of the fail-fast path that avoids parse_args for --help/-h and
--version/-V. This means cmd_line can be reused for the rest of the
lifetine of the program, and doesn't need to be in ArgStruct either.

After parse_args, nothing needs access to the "raw" args (just
cmd_line), so make args mutable and pass it as a Vec into parse_args.
parse_args can then pop flags and options from pargs until "--" is
reached, thus behaving the same as vanilla repo-add.