tidy/strip: Make strip and collecting of debug symbols in parallel
To be honest, it's far from a perfect implementation, but at least in my
opinion it's the most normal I can think of so far. It's pure bash, so
it's portable and at the same time gives some speed gain (about twice as
much on my machine when strip pacman, should tested more on larger
packages) and is pretty stable, I haven't seen any deviations from
normal behavior. This could be even faster if we could come up with a
mechanism where we could know exactly the current number of running
processes, although there is a ``wait -n`` in newer versions of Bash, my
experiments only resulted in race conditions, so it seems we can't do
without synchronizing mechanisms at least, but that complicates things a
lot in a script like this, so I don't think we need it now, a little
parallelization by “chunks” will suffice.

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>