tidy/zipkmod: Add KMOD_COMPRESSION_LEVEL to set the compression level
Hardcoding is not a good practice, even though this is used in zipman, I
think in the case of kernel modules we may want to change compression
level in the future.

We don't care if the user specifies an invalid value, because Zstd will
just ignore it and print warning.

Signed-off-by: Vasiliy Stelmachenok <ventureo@yandex.ru>