Partially revert resolving paths in pacman.conf
For the CacheDir and HookDir options, multiple paths can be specified
on a single line (separated by a space). The resolving of paths does
not handle multiple entries and causes failures. Additional, directories
in CacheDir are created by libalpm if they do not exist, and this is
not accounted for when realpath fails due to a non-exisiting directory.

"Fixes" #289 and #292.

Signed-off-by: Allan McRae <allan@archlinux.org>
(cherry picked from commit de68378f9c8e60963b76abdc770900b0e16a6cf4)